+++ tests/integration/test_session.py
@@ -1,7 +1,11 @@
+import os
+
 import pytest
 import requests
 
-from merqube_client_lib.session import _BaseAPISession
+from merqube_client_lib.constants import MERQ_REQUEST_ID_ENV_VAR
+from merqube_client_lib.exceptions import APIError
+from merqube_client_lib.session import MerqubeAPISession, _BaseAPISession
 
 
 @pytest.mark.parametrize("timeout_key", ["request_timeout", "timeout"], ids=["globally", "on-request"])
@@ -35,3 +39,21 @@ def test_timeout(caplog, timeout_key, timeout, delay, should_raise):
     else:
         response = sess.get(url)
         assert response.status_code == 200
+
+
+def test_request_id_generated():
+    sess = MerqubeAPISession()
+    with pytest.raises(APIError) as e:
+        sess.get("/api/v1/invalid-endpoint")
+    assert e.value.request_id is not None
+
+
+def test_request_id_user_passed():
+    sess = MerqubeAPISession()
+    try:
+        os.environ[MERQ_REQUEST_ID_ENV_VAR] = "ABCD-1234"
+        with pytest.raises(APIError) as e:
+            sess.get("/api/v1/invalid-endpoint")
+        assert e.value.request_id == "ABCD-1234"
+    finally:
+        del os.environ[MERQ_REQUEST_ID_ENV_VAR]